Handle sequential inserts with same position
The change introduced in #12434 was a fix for the case when a remove was
followed by an insert in the same position, which came in the wrong
order before #12434. This, however had the effect that sequential
inserts with the same position broke.

According to the spec, "If multiple inserts have the same position, the
order in the array defines the order in which the inserted strings
appear in the resulting text."

To fix both the case in #12434 and the spec case, this change keeps
track of any previous inserts and collapses them into a single change,
thus making sure that the inserted strings appear in the correct order.