jumplist: alleviate *, # command regression
Commit 35362495c ("jumplist: avoid extra tail entry #9805", 2019-04-02)
introduced a regression where, when the * or # command leaves the cursor
in the same line, the jumplist entry is erroneously erased. In addition
to interactive use, this breaks mappings like

    nnoremap * *<C-O>

The cursor remains in the same line when

1. the cursor is outside of a word boundary;
2. the searched-for word occurs next time in the same line; or
3. the searched-for word is unique.

Only erase the jumplist entry when both the cursor line and column
position are unchanged. This still leaves case #3 confusingly broken but
hopefully it's enough to salvage interactive use -- why "jump back" when
the cursor doesn't move? -- even if it turns out to be insufficient for
aforementioned mapping.

The slightly more clumsy alternative mapping of

    nnoremap * *``

doesn't help: `` updates the jumplist, too, so unique keywords move the
cursor to the position before the * command instead of staying in place.
If that command is used on a unique keyword after being used on a
non-unique keyword, that position is the "next match of the searched-for
word" from the previous usage.

ref #9874