Send buffer linecount in win_viewport
Currently, if GUIs want to implement scrollbars they'll have to query
the max line count of the current buffer for each window they want to
display the scrollbar for. Instead of requiring GUIs to do extra work to
get that one number, send it in the win_viewport event where all the
other related info is also send.

Ref #11748