ex_docmd: replace #define with enum
enum value can be inferred from previous member.
Vim was doing this manually via relative #define.

It's not needed but it is confusing for me to update an array index
after the array value and to configure the array index
such that it is dependent on other array indices.
One missing #define and everything below breaks.