Make open_floating_preview focusable
The preview window created by focusable_{float,preview} could be set to
be closed automatically as soon as a list of autocmds fired.
Unfortunately the autocommands WinLeave and BufLeave would fire before
leaving the current window, so it would be destroyed right before it'd
get focus.

This commit gives open_floating_preview the ability to be focusable
instead of needing to go through the wrapper functions focusable_preview
and focusable_float. The problematic autocmds BufLeave and WinLeave will
be filtered prior to focusing the preview window, and will be recreated
afterwards.

Additionally, the close_events are now customizable. We get that for
free, since open_floating_preview already needs to know about the list
of closing autocmds anyways, in order to implement focusing correctly.