Update popupmenu_spec.lua test for pum close
Patch 8.1.0035 has extra checks to see if insert mode ended from a
callback. If that is the case, it closes the popup menu.

The mouse-down test was edited to reflect this change:
Following the previous "split" test, the split is created with the popup
menu open. Following the mouse-down input_mouse function, the popup
window closes.

ref:
https://github.com/vim/vim/commit/6d41c78e353b630bc1a72cbff9160311d2a81e8c#diff-14dcf96f167b08fe1c89e75c9444b370R815-R821
https://github.com/neovim/neovim/blob/4042975df42152a64f4d390ed677013d1c8609c5/src/nvim/api/vim.c#L323
https://github.com/neovim/neovim/blob/e01fc33515d5cd3a9a595d0e6905bd16d3a861e5/src/nvim/os/input.c#L355