Update float_spec.lua test for pum close
Similar to popupmenu_spec.lua changes:

Patch 8.1.0035 has extra checks to see if insert mode ended from a
callback. If that is the case, it closes the popup menu.

When a new floating window buffer is created, it closes the pum and
exits insert mode. Add a screen comparison for this state, and adjust
followup tests by using feed 'a' instead of 'i' due to cursor moving
back -1 from the doESCkey

Also adjust some float_pos values to reflect the changes