help: sort help tag completions by name
When using help tag completion (`:h :syn-<tab>`), the matches looks unordered.
Especially when using wildoptions+=pum.

Actually, the result is sorted, but using a fancy (arbitrary) algorithm that
calculates a score for each tag name and and hides it in memory. Afterwards the
matches are sorted by these scores.

This commit removes the fancy algorithm and simply sorts tags by name.