Correct log levels
The log levels declared in the log module were inconsistent with the
MessageType object declared in the protocol module.

This surfaced as the inability to correctly log when using the
nvim-lspconfig package.

This change synchronizes the two logging levels so that the semantics
match, but does not make the naming consistent, as that would result in
a bigger change.