vim-patch:8.1.1717: last char in menu popup window highlighted
Problem:    Last char in menu popup window highlighted.
Solution:   Do not highlight an extra character twice.
https://github.com/vim/vim/commit/f914a33c9c8ec5c30da684a4a16edad3e0224f0a

N/A patches for version.c:

vim-patch:8.1.0746: highlighting not updated with conceal and 'cursorline'

Problem:    Highlighting not updated with conceal and 'cursorline'. (Jason
            Franklin)
Solution:   Do not use a zero line number.  Check if 'conceallevel' is set for
            the current window.
https://github.com/vim/vim/commit/bbee8d5122b159683b3f52eddd0da85fcf1fcbfd