doc: Document CursorMoved vim/neovim difference
In Vim, the CursorMoved event is triggered either when the cursor is
moved in Normal or Visual mode or the text of the cursor line is
changed. In Neovim, the CursorMoved event is also triggered when the
cursor is moved to another window. This commit documents this nuance
in runtime/doc/vim_diff.txt.