Truncate LSP diagnostic virtual text on new line.
Currently when we have a diagnostic that contains a new line, we simply
strip it out and replace it with a space. For some languages this may
work fine, but for a lot of typed languages that have longer signatures
or types and also multiline diagnostics, this falls apart pretty quickly.
This takes a different approach and if a new line is detected it will
instead trunacate the diagnostic message (only for the virtual text) and
only display the first line making a more succinct message. Then when they
go to view the diagnostic they will continue to see the full message as they
would have before.