Fix excessive grid_scrolls on ext_messages+small window
This commit fixes excessive redraws that would happen when ext_messages
is used with a scrolloffset larger than the window.

The reason we need to use `plines <= so * 2` is that the window needs to
accomodate both the scroll offset (so * 2) and the line the cursor is
on, so there isn't enough room even if plines is equal to (so * 2).

Closes #12810.