fix(lsp): stop loclist from opening when empty
Currently using the `vim.lsp.diagnostic.set_loclist({open_loclist = true})`
function opens the location list even if it is empty. This is mostly
fine if youre only using the function as a callback and checking if
there are error items beforehand. However, if we want to tie this
function into a autocmd, like perhaps `LspDiagnosticChanged`, then this
behavior becomes quite noisey. It forces the location list open even
when empty, and becomes the focused window. This makes this option
virtually unusuable when in conjunction with the autocmd – and from a
quick google search, appears to be atleast somewhat common.

So then the solution could be to use `lwindow` instead of the
needy `lopen' command. This way, we only open the location list when it
is of interest to the user, when it has items to display! Now thats
smooth.