ci(GHA/release): Delete the release when creating a new nightly
This reverts #13733.  Disabling delete of the release leaves behind a
draft release associated with the previous nightly tag, which is
annoying to cleanup.

In the rare circumstances that the creation of a new nightly fails after
deleting the previous nightly, the job can be re-run or a nightly can be
missing for a day.