win/package: remove cat,tidy
Neovim should not bundle external tools
that are not needed in the runtime environment.

cat.exe is meant for tests only.
Install a mingw/msys2/busybox environment which bundle cat.exe.
tidy.exe was never used in tests and is not required in Neovim runtime.
busybox and tidy.exe can be installed via scoop.

Ref: https://github.com/neovim/neovim/issues/14078