Extract `diff_parse` into a pure function
This commit refactors `diff_parse` out of `diff_read`. This should
improve readable and allow for reuse of diff file parsing in other parts
of the codebase. In particular, we can now reuse `diff_parse` in
`run_external_diff_on_sample`.

An additional benefit of this refactoring is that `diff_parse` can be
and is unit-tested.