fix(lsp): Make focusable_{preview,float} focusable
The preview window created by focusable_{float,preview} could be set to
be closed automatically as soon as a list of autocmds fired.
However, the autocommands WinLeave and BufLeave fire before leaving the
'parent' window, so the preview window would be destroyed right before
it'd get focus, throwing an error instead.

This is just the tiny one-line fix from the spaghetti mess that #12720
has become.