fix(lsp): runtime error buf_request_all if no server supports a request
When no server would be active to handle the request, `_sync_handler`
would fail because the `client_id` is `nil`. Furthermore
`set_expected_result_count` fails because `client_request_ids` is `nil`.
This commit adds safety guards for this case, to return an empty mapping
instead of failing with a runtime error.