fix!: make :runtime work as advertised
With no [where] argument, the `:runtime` command should only execute
files directly under the user's 'runtimepath'. However, the flags are
being modified to also search under the start directory within
'packpath' when there is no [where] argument. This breaks the contract
of how :runtime is supposed to work according to the documentation and
results in files being sourced when they shouldn't be.

The motivation behind modifying the flags was to allow users to access
files under start directories from their init.vim or init.lua files;
that is, _before_ packages had been loaded and added to the
'runtimepath'. However, Neovim already provides a mechanism to do this
that aligns with documented and established behavior through the use of
:packload and :packloadall.

Revert the flag modification to have :runtime again work as documented.