feat(api): allow arguments to be optional
Currently lua functions in vim.api have an argument length check. This
check is pointless since the users can fake the arguments by passing in
nils. E.g.

  vim.api.nvim_win_get_width()
  > E5108: Error executing lua [string ":lua"]:1: Expected 1 argument

However:

  vim.api.nvim_win_get_width(nil)
  > 98

This change allows API function arguments to be omitted by pushing extra
nils onto the stack.

The main benefit to this change is that it allows extra arguments to be
added to functions without introducing a breaking change.