feat(treesitter): support query quantifiers
Do this by making tables instead of just nodes for when quantifiers are
used, and match multiple nodes.

Also get away from set_match and use push_match, a function that is
responsible of creating and pushing a match table onto the stack.
Also stop reusing the same table for the matches in iter_captures, even
if that causes reallocation.